ll: Replace constant address with variables
There is a single global variable which is used as base address for all
accesses to absolute memory locations. It is initialized with the first
constant memory address found. In the JIT (ExecutionEngine) it is again
bound to the original location, so that the functionality is identical
but the LLVM IR does not contain addresses any more.