ll: Implement heuristic to guess value type
When doing an arithmetic operation, we can improve the code by using
pointer arithmetics. However, if the value type is actually an integer,
the result will be worse. So, we need to guess whether a value is a
pointer or not just by looking at the value in the current basic block.