ll: Pass alignment information to LLVM
This allows an instruction to specify the maximum alignment of memory
operands based on the operand size or a specific lower alignment.

However, x86 does only enforce alignment for SSE instructions but not
for general purpose instructions (like MOV or ADD). For these, the
alignment will be still set to maximum as we can expect that LLVM emits
general purpose instructions again for these memory operands.