Start port to static-only emulation
This puts a wrapper around existing emulation,
and introduces an processing context, which
is filled with in/out behavior of an instruction.

If that is provided, there also must be separate
handlers for capturing and emulation, to be used
for the different cases (dynamic/static).

This allows for instruction-wise porting to new
handlers.

Do this for cltq/cwtl, with test. Works.